Go live code

+++ angular-app/src/app/pages/cart/cart.component.ts
@@ -15,7 +15,7 @@ export class CartComponent implements OnInit {
 
   mobileNumberEntered = false;
   enterCoupon = false;
-  cart : any;
+  cart : any = {};
   sessionCheckInterval : any;
   cartOpen = false;
   mobileNumber : any;
@@ -31,7 +31,10 @@ export class CartComponent implements OnInit {
   }
   reloadSubscription: Subscription;
   loadSubscription: Subscription;
+  closeModalSubscription: Subscription;
+  openModalSubscription : Subscription;
   cartItemOutOfStock : boolean = false;
+  fetchCartFailed : boolean = false;
   constructor( private router: Router,
                private appservice : AppServiceService,
                private apiservice : ApiServiceService,
@@ -39,9 +42,14 @@ export class CartComponent implements OnInit {
               ) { 
     this.reloadSubscription = this.appservice.listenToAddToCartEvent().subscribe(()=> { this.reloadCart() });
     this.loadSubscription = this.appservice.listenToOpenCartEvent().subscribe(()=> { this.loadCart() });
+
+    this.closeModalSubscription = this.appservice.listenToCloseModal().subscribe(()=>{ this.updateOtpModal(false)});
+    this.openModalSubscription = this.appservice.listenToOpenModal().subscribe(()=>{ this.modalHandler()});
+
   }
 
   reloadCart(){
+    this.updateUrl();
     console.log("listened to the add to cart trigger");
     this.cartOpen = true;
     this.fetchCartDataOnAddToCartSuccess();
@@ -49,6 +57,7 @@ export class CartComponent implements OnInit {
   }
 
   loadCart(){
+    this.updateUrl();
     console.log("listened to open cart trigger");
     this.cartOpen = true;
     this.getCartData();
@@ -57,37 +66,49 @@ export class CartComponent implements OnInit {
   ngOnDestroy() {
   // unsubscribe to ensure no memory leaks
     this.reloadSubscription.unsubscribe();
-    this.loadSubscription.unsubscribe()
+    this.loadSubscription.unsubscribe();
+    this.closeModalSubscription.unsubscribe();
+    this.openModalSubscription.unsubscribe();
   }
 
   ngOnInit() {
-    this.cartOpen = true;
-    $('.ng-cart-loader').removeClass('cart-loader')
-    if(sessionStorage.getItem('add_to_cart_clicked')){
-      console.log("add to cart clicked");
-      this.fetchCartDataOnAddToCartSuccess();
-      sessionStorage.removeItem('add_to_cart_clicked');
-    }
-    else{
-      this.getCartData();
-    }
+    console.log("ngOnInit cart component");        
+      this.updateUrl();
+      this.cartOpen = true;
+      $('.ng-cart-loader').removeClass('cart-loader')
+      if(sessionStorage.getItem('add_to_cart_clicked')){
+        this.fetchCartDataOnAddToCartSuccess();
+        sessionStorage.removeItem('add_to_cart_clicked');
+      }
+      else{
+        this.getCartData();
+      }
+  }
+
+  ngAfterViewInit(){
+    if(window.location.href.endsWith('/user-verification'))
+      this.modalHandler();
+  }
+
+  updateUrl(){
+    let url = window.location.href.split("#")[0] + '#bag';
+    if(window.location.href.endsWith('#shipping-address') || window.location.href.endsWith('#shipping-summary'))
+      history.replaceState({cart : true}, 'cart', url);
+    else if(!(window.location.href.endsWith('#bag') || window.location.href.endsWith('#bag/user-verification')))
+      history.pushState({cart : true}, 'cart', url);
   }
 
   fetchCartDataOnAddToCartSuccess(){    
     this.appservice.showLoader()
     if(sessionStorage.getItem('cart_data')){
       this.cart = JSON.parse(sessionStorage.getItem('cart_data'));
-      console.log("cart_data from sessionStorage==>", this.cart);
     }
     else
-      this.cart = { items : [] };
+      this.cart = {};
     
     this.sessionCheckInterval = setInterval(()=>{
       if(sessionStorage.getItem('addded_to_cart')){
-        if(sessionStorage.getItem('addded_to_cart') == "true")
-          this.fetchCartDataFromServer();
-        else
-          this.appservice.removeLoader()
+        this.fetchCartDataFromServer();
         sessionStorage.removeItem('addded_to_cart');
         clearInterval(this.sessionCheckInterval);
       }
@@ -100,55 +121,39 @@ export class CartComponent implements OnInit {
     this.appservice.showLoader()
     if(sessionStorage.getItem('cart_data')){
       this.cart = JSON.parse(sessionStorage.getItem('cart_data'));
-      console.log("cart_data from sessionStorage==>", this.cart);
     }
     this.fetchCartDataFromServer();
   }
 
   fetchCartDataFromServer(){
     this.appservice.showLoader()
-    let url = this.appservice.apiUrl + (this.isLoggedInUser() ? ("/api/rest/v1/user/cart/"+this.appservice.getCookie('cart_id')+"/get") : ("/rest/anonymous/cart/get"))
-    console.log(this.isLoggedInUser());
-    let header = this.isLoggedInUser() ? { Authorization : 'Bearer '+this.appservice.getCookie('token') } : {}
-    this.apiservice.request(url, 'get', {}, header ).then((response)=>{
-      console.log("response ==>", response);
-      this.cart = this.calculateOffPercenatge(response);
+    this.appservice.callFetchCartApi().then((response)=>{
+      this.cart = this.formattedCartDataForUI(response);
       this.checkCartItemOutOfStock();
-      sessionStorage.setItem('cart_data', JSON.stringify(this.cart));
-      document.cookie = "cart_count=" + this.cart.cart_count + ";path=/";
-      this.appservice.updateCartCountInUI();
-      this.appservice.removeLoader()
+      this.updateLocalDataAndUI(this.cart, this.cart.cart_count);
+      this.appservice.removeLoader();
+      this.fetchCartFailed = false;
       this.zone.run(() => {});
     })
     .catch((error)=>{
-      console.log("error ===>", error);
-      // if(error.message == "Cart not found for this session"){
-        this.cart = {
-          items : []
-        }
-      // }
-      this.appservice.removeLoader()
+      this.handleFetchCartFailure(error);
+      this.appservice.removeLoader();
       this.zone.run(() => {});
     })
     this.zone.run(() => {});
   }
 
-  calculateOffPercenatge(data){
+  formattedCartDataForUI(data){
     data.items.forEach((item)=>{
       if(item.attributes.price_mrp != item.attributes.price_final)
         item.off_percentage = Math.round(((item.attributes.price_mrp - item.attributes.price_final) / (item.attributes.price_mrp )) * 100) + '% OFF';
       item.href = '/' + item.product_slug +'/buy?size='+item.attributes.size;
+      item.attributes.images = Array.isArray(item.attributes.images) ? ['/img/placeholder.svg', '/img/placeholder.svg', '/img/placeholder.svg'] : Object.values(item.attributes.images);
     })
 
     data.items.sort((a,b)=>{ return a.timestamp - b.timestamp});
     data.items.reverse();
     return data;
-  }
-
-  isLoggedInUser(){
-    if(this.appservice.getCookie('token') && this.appservice.getCookie('cart_id'))
-      return true;
-    return false;
   }
 
   modifyCart(item){
@@ -160,7 +165,7 @@ export class CartComponent implements OnInit {
     //   quantity : item.quantity
     // }
     // console.log("Body ==>", body);
-    // let url = 'http://localhost:8000/rest/anonymous/cart/update';
+    // let url = 'http://localhost:8000/rest/v1/anonymous/cart/update';
     // this.apiservice.request(url, 'get', body ).then((response)=>{
     //   console.log("response ==>", response);
     //   item = response.item;
@@ -173,26 +178,31 @@ export class CartComponent implements OnInit {
 
   deleteItem(item){
     this.appservice.showLoader()
-    console.log("delete item ==>", item);
     let body = { variant_id : item.id };
-    let url = this.appservice.apiUrl + (this.isLoggedInUser() ? ("/api/rest/v1/user/cart/"+this.appservice.getCookie('cart_id')+"/delete?") : ("/rest/anonymous/cart/delete?"));
-    let header = this.isLoggedInUser() ? { Authorization : 'Bearer '+this.appservice.getCookie('token') } : {};
+    let url = this.appservice.apiUrl + (this.appservice.isLoggedInUser() ? ("/api/rest/v1/user/cart/"+this.appservice.getCookie('cart_id')+"/delete?") : ("/rest/v1/anonymous/cart/delete?"));
+    let header = this.appservice.isLoggedInUser() ? { Authorization : 'Bearer '+this.appservice.getCookie('token') } : {};
     url = url+$.param(body);
     this.apiservice.request(url, 'get', body, header ).then((response)=>{
-      console.log("response ==>", response);
       let index = this.cart.items.findIndex(i => i.id == item.id)
       this.cart.items.splice(index,1);
       this.cart.summary = response.summary;
       this.cart.cart_count = response.cart_count;
       this.checkCartItemOutOfStock();
-      document.cookie = "cart_count=" + this.cart.cart_count + ";path=/";
-      sessionStorage.setItem('cart_data', JSON.stringify(this.cart));
-      this.appservice.updateCartCountInUI();
+      this.updateLocalDataAndUI(this.cart, this.cart.cart_count);
       this.appservice.removeLoader()
     })
     .catch((error)=>{
       console.log("error ===>", error);
-      this.appservice.removeLoader()
+      if(error.status == 401){
+        this.appservice.userLogout();
+        this.fetchCartDataFromServer();
+        this.fetchCartFailed = false; 
+      }
+      else if((error.status == 400 || error.status == 403) && this.appservice.isLoggedInUser() ){
+        this.getNewCartId();
+        this.fetchCartFailed = false; 
+      }
+      this.appservice.removeLoader();
     })
   }
 
@@ -205,14 +215,10 @@ export class CartComponent implements OnInit {
   authenticateUser(){
     this.userValidation.disableSendOtpButton = true;
     this.userValidation.mobileValidationFailed = false;
-    // this.appservice.apiUrl = 'http://demo8558685.mockable.io/';
     let url = this.appservice.apiUrl + '/rest/v1/authenticate/generate_otp?';
-    let body = {
-      phone : this.mobileNumber
-    }
+    let body = { phone : this.mobileNumber };
     url = url+$.param(body);
     this.apiservice.request(url, 'get', body , {}, true).then((response)=>{
-      console.log("response ==>", response);
       this.userValidation.disableSendOtpButton = false;
       if(response.success){
         this.mobileNumberEntered = true;
@@ -221,7 +227,6 @@ export class CartComponent implements OnInit {
         this.userValidation.mobileValidationFailed = true;
         this.userValidation.mobileValidationErrorMsg = response.message
       }
-
     })
     .catch((error)=>{
       console.log("error ===>", error);
@@ -235,19 +240,20 @@ export class CartComponent implements OnInit {
     this.userValidation.otpVerificationFailed = false;
     let url = this.appservice.apiUrl + '/rest/v1/authenticate/login?';
     // this.otp=this.otpCode.otp1+this.otpCode.otp2+this.otpCode.otp3+this.otpCode.otp4+this.otpCode.otp5+this.otpCode.otp6
-    console.log("OTP ==>", this.otp);
     let body = {
       otp : this.otp,
       phone : this.mobileNumber
     }
     url = url + $.param(body);
     this.apiservice.request(url, 'get', body, {}, true).then((response)=>{
-      console.log("response ==>", response);
       this.otp = null;
       this.userValidation.disableVerifyOtpButton = false;
       if(response.success){
         document.cookie='token='+ response.token + ";path=/";
         document.cookie='cart_id=' + response.user.active_cart_id + ";path=/";
+        this.appservice.userVerificationComplete = true;
+        // this.updateOtpModal(false);
+        $('body').removeClass('modal-open')
         this.navigateToShippingDetailsPage();        
       }
       else{
@@ -260,22 +266,23 @@ export class CartComponent implements OnInit {
       console.log("error ===>", error);
       this.userValidation.disableVerifyOtpButton = false;
       this.userValidation.otpVerificationFailed = true;
-    })
-  	
+    })  	
   }
 
   closeCart(){
-    // this.cart = null;
-    this.cartOpen = false;
+    let url = window.location.href.split("#")[0];
+    history.pushState({cart : false}, 'cart', url);
     this.appservice.closeCart();
   }
 
   modalHandler(){
-    console.log("modalHandler")
-    if(this.isLoggedInUser()){
+    if(this.appservice.isLoggedInUser()){
       this.navigateToShippingDetailsPage();
     }
     else{
+      let url = window.location.href +'/user-verification';
+      if(!window.location.href.endsWith('#bag/user-verification'))
+        history.pushState({cart : true}, 'cart', url);
       $('#signin').modal('show');
       $("#cd-cart").css("overflow", "hidden");
       $('.modal-backdrop').appendTo('#cd-cart');
@@ -295,7 +302,7 @@ export class CartComponent implements OnInit {
       return true;
   }
 
-  updateOtpModal(){
+  updateOtpModal(updateHistory : boolean = true){
     $('#signin').modal('hide');
     $("#cd-cart").css("overflow", "auto");
     this.mobileNumberEntered = false;
@@ -303,23 +310,29 @@ export class CartComponent implements OnInit {
     this.userValidation.otpVerificationErrorMsg = '';
   }
 
+  closeOtpModal(){
+    history.back();
+  }
+
   navigateToShippingDetailsPage(){
-    this.appservice.showLoader();
-    let url = this.appservice.apiUrl + "/api/rest/v1/user/address/all";
-    console.log(this.isLoggedInUser());
-    let header = this.isLoggedInUser() ? { Authorization : 'Bearer '+this.appservice.getCookie('token') } : {};
-    this.apiservice.request(url, 'get', {} , header ).then((response)=>{
-      console.log("response ==>", response);
-      this.appservice.shippingAddresses = response.addresses;
-      $("#cd-cart").css("overflow", "auto");
-      $('.modal-backdrop').remove();
-      this.router.navigateByUrl('/shipping-details', { skipLocationChange: true });
-      this.appservice.removeLoader();
-    })
-    .catch((error)=>{
-      console.log("error ===>", error);
-      this.appservice.removeLoader();
-    })
+    if(this.cart.cart_type == "cart"){
+      this.appservice.showLoader();
+      this.appservice.callGetAllAddressesApi().then((response)=>{
+        this.appservice.shippingAddresses = response.addresses;
+        $("#cd-cart").css("overflow", "auto");
+        $('.modal-backdrop').remove();
+        this.router.navigateByUrl('/shipping-details', { skipLocationChange: true });
+        this.appservice.removeLoader();
+      })
+      .catch((error)=>{
+        console.log("error ===>", error);
+        this.appservice.removeLoader();
+      })
+    }
+    else{
+      this.router.navigateByUrl('/shipping-summary', { skipLocationChange: true });
+      this.appservice.continueOrder = true;
+    }
   }
 
   checkCartItemOutOfStock(){
@@ -330,6 +343,79 @@ export class CartComponent implements OnInit {
         // break;
       }
     })
+  }
+
+  editBag(){
+    this.appservice.showLoader();
+    let url = this.appservice.apiUrl + '/api/rest/v1/user/cart/start-fresh';
+    let header = { Authorization : 'Bearer '+this.appservice.getCookie('token') };
+    this.apiservice.request(url, 'get', {} , header ).then((response)=>{
+      document.cookie='cart_id=' + response.cart_id + ";path=/";
+      this.fetchCartDataFromServer();
+    })
+    .catch((error)=>{
+      console.log("error ===>", error);
+      this.appservice.removeLoader();
+    })
+  }
+
+  reloadPage(){
+    window.location.reload();
+  }
+
+  getNewCartId(){
+    this.appservice.showLoader();
+    this.appservice.callMineApi().then((response)=>{
+      if(response.cart_id != this.appservice.getCookie('cart_id')){
+        document.cookie='cart_id=' + response.cart_id + ";path=/";
+        this.fetchCartDataFromServer();
+      }
+      else{
+        this.cart = {};
+        this.fetchCartFailed = true;
+      }
+    })
+    .catch((error)=>{
+      console.log("error : mine api===>", error);
+      this.appservice.removeLoader();
+    })
+  }
+
+  shopNow(){
+    window.location.href = '/shop'
+  }
+
+  updateLocalDataAndUI(cart : any = null, cart_count = 0){
+    if(cart)
+      sessionStorage.setItem('cart_data', JSON.stringify(cart));
+    else
+      this.appservice.clearSessionStorage();
+    document.cookie = "cart_count=" + cart_count + ";path=/";
+    this.appservice.updateCartCountInUI();
+  }
+
+  handleFetchCartFailure(error){
+    if(error.status == 401){
+      this.appservice.userLogout();
+      this.fetchCartDataFromServer();
+      this.fetchCartFailed = false; 
+    }
+    else if((error.status == 400 || error.status == 403) && this.appservice.isLoggedInUser() ){
+      this.getNewCartId();
+      this.fetchCartFailed = false; 
+    }
+    else if(error.status == 404){
+      this.cart = {
+        items : []
+      }
+      this.fetchCartFailed = false;
+      this.updateLocalDataAndUI();
+    }
+    else{
+      this.cart = {};
+      this.fetchCartFailed = true;
+      this.updateLocalDataAndUI();
+    }
   }
   
 }
